test: split RecordingManager suite into thematic files with a shared kit
The 3400-line RecordingManager.test.ts becomes five thematic suites -
lifecycle (21), streaming/write-chain/auto-split (24), output and note
insertion (17), session journal (6), and marker drafts (9) - plus a
shared tests/unit/helpers/recordingManagerTestKit.ts carrying the
media stubs, mock app factory, recorder double, and marker-store
fakes that each suite previously duplicated. The fallback suite now
uses the kit too instead of its byte-identical stub block. Test bodies
are copied verbatim; total count is unchanged (86).

One correction to the audit's claim that the two createDesktopRecorder
helpers were identical: the write-chain variant reset Platform to
desktop and the journal variant did not (the mobile-session journal
test depends on that). The kit keeps the non-resetting variant and the
write-chain suite restores desktop platform in its own beforeEach.
